+++ SampleProject/Core/Services/Users/DeleteUserService.cs
@@ -1,10 +1,12 @@
 ﻿using BusinessEntities;
 using Common;
+using Common.Results;
 using Data.Repositories;
+using System;
 
 namespace Core.Services.Users
 {
-    [AutoRegister]
+    [AutoRegister(AutoRegisterTypes.Scope)]
     public class DeleteUserService : IDeleteUserService
     {
         private readonly IUserRepository _userRepository;
@@ -14,9 +16,23 @@ namespace Core.Services.Users
             _userRepository = userRepository;
         }
 
-        public void Delete(User user)
+        public Result<User> DeleteById(Guid id)
         {
+            var user = _userRepository.Get(id);
+            if (user == null)
+                return Result<User>.Fail("User not found.");
+
             _userRepository.Delete(user);
+            return Result<User>.Ok(user);
+        }
+
+        public Result<User> Delete(User user)
+        {
+            if (user == null)
+                return Result<User>.Fail("User cannot be null.");
+
+            _userRepository.Delete(user);
+            return Result<User>.Ok(user);
         }
 
         public void DeleteAll()
